+++ src/main/java/io/khasang/genelove/dao/UserDAO.java
@@ -14,7 +14,7 @@ public interface UserDAO {
 
     User getUserByLogin(String login);
 
-    User getUserById(int id);
+    User getUserById(long id);
 
     List<User> getUserAll();
 
@@ -22,7 +22,7 @@ public interface UserDAO {
 
     Role getRoleByName(String name);
 
-    Role getRoleById(int id);
+    Role getRoleById(long id);
 
     void addFavourite(User user, User favourite);
